+++ PyPWA/core/shared/interfaces/internals.py
@@ -145,16 +145,6 @@ class ProcessInterface(object):
         """
         raise NotImplementedError
 
-    @property
-    def previous_value(self):
-        # type: () -> Any
-        """
-        The previous value received from the kernel interface.
-
-        :return: Previous calculated value.
-        """
-        raise NotImplementedError
-
     def stop(self, force=False):
         # type: (Opt[bool]) -> None
         """
@@ -208,8 +198,8 @@ class Kernel(object):
 class KernelInterface(object):
 
     # is_duplex controls whether the kernel will shutdown after its first
-    # run or if the kernel will wait for more information. If this is true
-    # the kernel will only be able to receive data and will shutdown after
+    # run or if the kernel will wait for more information. If this is false
+    # the kernel will only be able to send data and will shutdown after
     # its first run; however, if this is true then the kernel will stay
     # running and waiting for a value from the interface indefinitely until
     # they are shutdown manually.
@@ -251,3 +241,9 @@ class LikelihoodTypes(enum.Enum):
     OTHER = 1
     CHI_SQUARED = 2
     LOG_LIKELIHOOD = 3
+
+
+class ProcessCodes(enum.Enum):
+
+    SHUTDOWN = 1
+    ERROR = 2